+++ com_simplestats/admin/sql/uninstall.mysql.utf8.sql
@@ -1,1 +1,4 @@
+DROP TABLE IF EXISTS `#__simplestats_daily_items`;
+DROP TABLE IF EXISTS `#__simplestats_daily_dimensions`;
+DROP TABLE IF EXISTS `#__simplestats_daily`;
 DROP TABLE IF EXISTS `#__simplestats_events`;